user_management/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
djangorestframework==3.15.2
sqlparse==0.5.1
gunicorn==23.0.0
psycopg2-binary==2.9.9

